Refactor Contact container
- Move contact containner to ./containers/Contacts
