+++ src/containers/Contacts/Contact.test.js
@@ -1,5 +1,5 @@
 import React from "react";
-import Contact from "./ContactContainer";
+import Contact from "./Contacts";
 import {
   shallow
 } from 'enzyme'
